requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
openai-whisper
torch>=2.2.0
torchaudio>=2.2.0
numpy>=1.24.0
setuptools>=65.0.0
gunicorn==21.2.0
ffmpeg-python==0.2.0
